requirements.txt
python-dotenv>=1.0.1
langchain>=0.1.16
langchain_ollama>=0.1.1
langchain_core>=0.1.32
pandas>=1.5.3
nltk>=3.8.1
rouge_score>=0.1.2
sacrebleu>=2.3.1
unbabel-comet>=2.0.0
torch==2.1.2
transformers==4.37.2
bert-score 
bleurt-pytorch